Decouple UI and adjust some actions
