+++ podcasts/en.lproj/Localizable.strings
@@ -2028,6 +2028,9 @@
 /* Episodes will be displayed in order from the shortest to the longest. */
 "podcasts_episode_sort_shortest_to_longest" = "Shortest to Longest";
 
+/* Episodes will be displayed in custom order by drag and drop */
+"podcasts_episode_sort_drag_and_drop" = "Drag and Drop";
+
 /* Presents the podcasts with large podcast artwork tiles. */
 "podcasts_large_grid" = "Large Grid";
